refactor CLI dir
